Modularização dos Models e correção de sintaxe
