+++ registry/models/__init__.py
@@ -1,0 +1,4 @@
+from .contact import Contact
+from .municipality import Municipality
+from .person import Person, IndividualPerson, BusinessPerson
+from .address import Address
